refactor(haven): address CodeRabbit review on interactive prune
- DiskUsage: keep du's grand total when it exits non-zero on an
  unreadable subdir (it still prints a usable number); the verdict is now
  "did it print a number", not "did it exit 0".
- parseNonNegInt: use strconv.Atoi so "5abc" is rejected rather than
  silently read as 5 (fmt.Sscanf ignored trailing garbage).
- Dedup: add domain.SlugOrBase + domain.DBChips; the picker and the report
  both call them instead of each carrying its own copy.
- Split the 918-line prunetui.go by responsibility into prunetui.go
  (surface + Update), model.go (state + delete), order.go (sort + scroll),
  view.go (render), format.go (helpers) — each now well under 300 lines.

(The unbound-scenario finding was already fixed in the prior commit.)
